+++ src/data/issues/kismet.js
@@ -1,0 +1,94 @@
+export const kismetItems = [
+    {
+        id: "kismet-1",
+        title: "The Lady Of The Castle",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/the%20lady%20of%20the%20castle/kismetnew.jpg",
+        to: "/legacy/articles/kismet/theladyofthecastle.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-2",
+        title: "Judas",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/judas/illustration.jpeg",
+        to: "/legacy/articles/kismet/judas.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-3",
+        title: "Ephemeral",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/ephemeral/Group6_KISMET_Graphics4.jpeg",
+        to: "/legacy/articles/kismet/ephemeral.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-4",
+        title: "Give me a sign",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/give%20me%20a%20sign/GMA_p2.png",
+        to: "/legacy/articles/kismet/giveMeASign.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-5",
+        title: "Saboteur",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/saboteur/thirdPic.png",
+        to: "/legacy/articles/kismet/saboteur.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-6",
+        title: "What Remains Of A Star",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/meteor%20shower%3Awhat%20remains%20of%20a%20star/kismet%20-%20Shravya%20(1)%20image%20edit%202-min.png",
+        to: "/legacy/articles/kismet/whatRemainsOfAStar.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-7",
+        title: "Meteor Shower",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/meteor%20shower%3Awhat%20remains%20of%20a%20star/darrenCoverPic.png",
+        to: "/legacy/articles/kismet/meteorShower.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-8",
+        title: "I Know The End",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/i%20know%20the%20end/KISMET_G3_1.jpg",
+        to: "/legacy/articles/kismet/iknowtheend.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-9",
+        title: "किस्मत (Kismet)",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/metamorphosis%20(kismet)/be_picture23.jpg",
+        to: "/legacy/articles/kismet/kismetarticle.html",
+        coverPos: "center center",
+    },
+    {
+        id: "kismet-10",
+        title: "Metamorphosis",
+        description: "Kismet Issue",
+        tags: ["KISMET"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/9%20-%20kismet/metamorphosis%20(kismet)/be_picture10.jpg",
+        to: "/legacy/articles/kismet/metamorphosis.html",
+        coverPos: "center center",
+    },
+];
+
+export default kismetItems;
