+++ src/data/issues/not.js
@@ -1,0 +1,58 @@
+export const notItems = [
+    {
+        id: "not-1",
+        title: "Echosystems",
+        description: "Writer: Sanhita P.",
+        tags: ["NOT"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/10%20-%20NOT/ecosystems/group2coverpic.jpg",
+        to: "/legacy/articles/NOT/echosystems.html",
+        coverPos: "center center",
+    },
+    {
+        id: "not-2",
+        title: "In the Wind",
+        description: "Writer: Sarah R.",
+        tags: ["NOT"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/10%20-%20NOT/in%20the%20wind/Sarah_Rizvi_Final_Graphic_3.jpeg",
+        to: "/legacy/articles/NOT/inthewind.html",
+        coverPos: "center center",
+    },
+    {
+        id: "not-3",
+        title: "Quiet Devourment",
+        description: "Writer: Emma L.",
+        tags: ["NOT"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/10%20-%20NOT/quiet%20devourment/cover.jpg",
+        to: "/legacy/articles/NOT/quietdevourment.html",
+        coverPos: "center center",
+    },
+    {
+        id: "not-4",
+        title: "La Quinta",
+        description: "Writer: Zayeed A.",
+        tags: ["NOT"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/10%20-%20NOT/la%20quinta/LQcover.png",
+        to: "/legacy/articles/NOT/laQuinta.html",
+        coverPos: "center center",
+    },
+    {
+        id: "not-5",
+        title: "Air It Out",
+        description: "Writer: Susan E.",
+        tags: ["NOT"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/10%20-%20NOT/air%20it%20out/3.jpg",
+        to: "/legacy/articles/NOT/airitout.html",
+        coverPos: "center center",
+    },
+    {
+        id: "not-6",
+        title: "Disconnect",
+        description: "Writer: Vivika I.",
+        tags: ["NOT"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/10%20-%20NOT/disconnect/rojanLaying.JPG",
+        to: "/legacy/articles/NOT/Disconnect.html",
+        coverPos: "center center",
+    },
+];
+
+export default notItems;
