+++ src/data/issues/serial.js
@@ -1,0 +1,96 @@
+export const serialItems = [
+    {
+        id: "article10",
+        title: "Life in Paradise",
+        description: "By: Kerstin T. & Manjiri C.",
+        tags: ["INTERACTIVE", "VISUAL", "MAGAZINE"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/life%20in%20paradise/group9cover.png",
+        to: "/issues/serial/lifeinparadise",
+        coverPos: "center center",
+    },
+    {
+        id: "article-7",
+        title: "ctrl + c",
+        description: "By: Zenah I.",
+        tags: ["IMMERSIVE"],
+        image: "https://res.cloudinary.com/dyj46skdc/image/upload/v1765339839/facestretch_sadc29.png",
+        to: "/issues/serial/indigoos",
+        coverPos: " left 25% top 10%",
+    },
+    {
+        id: "article-5",
+        title: "Stomach Ache",
+        description: "By: Sanhita Perkari",
+        tags: ["INTERACTIVE"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/stomachache/serialpicture.JPG",
+        to: "/issues/serial/stomachache",
+        coverPos: "center center",
+    },
+    {
+        id: "article-4",
+        title: "Vengeance of the West",
+        description: "By: Karishma Pilla",
+        tags: ["Western", "Serial Killer"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/vengeance%20of%20the%20west/teaser.jpg",
+        to: "/issues/serial/western",
+        coverPos: "center center",
+    },
+    {
+        id: "article-9",
+        title: "Covet",
+        description: "By: Joanna Virippil",
+        tags: ["Hate", "Friendship"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/covet/9-clothes-2.jpg",
+        to: "/issues/serial/covet",
+        coverPos: "center center",
+    },
+    {
+        id: "article-8",
+        title: "You Can Twist Time",
+        description: "By: Jiya Gupta",
+        tags: ["SCROLLABLE", "NOVEMBER"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/you%20can%20twist%20time/final.png",
+        to: "/issues/serial/youcantwisttime",
+        coverPos: "center center",
+    },
+    {
+        id: "article-3",
+        title: "Keyboards",
+        description: "By: Manogna Bedhu",
+        tags: ["INTERACTIVE", "NOVEMBER"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/keyboards/anotha%20one.png",
+        to: "/issues/serial/keyboards",
+        coverPos: "center center",
+    },
+    {
+        id: "article-2",
+        title: "A Number Out of Place ",
+        description: "By: Nicole C. ",
+        tags: ["MATRIX", "NOVEMBER"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/a%20number%20out%20of%20place/REN_1404%20copy.jpg",
+        to: "/issues/serial/anumberoutofplace",
+        coverPos: " left 25% top 10%",
+    },
+    {
+        id: "article-1",
+        title: "I Love Shopping ",
+        description: "By: Zayeed A. ",
+        tags: ["FASHION", "NOVEMBER"],
+        image: "https://cdn.indigomagazinetx.com/article%20photos/11%20-%20serial/i%20love%20shopping/BF5T7684.jpg",
+        to: "/issues/serial/iloveshopping",
+        coverPos: " left 25% top 10%",
+    },
+    {
+        id: "commingSoon",
+        description: "",
+        image: "/legacy/assets/backgrounds/More.png",
+        imageGrid: "/legacy/assets/backgrounds/dott.svg",
+        gridPos: "top 50% left 50%",
+        title: "More Coming Soon",
+        to: "",
+        blur: true,
+        comingSoon: true,
+    },
+];
+
+export default serialItems;
