--- src/data/posts/covet.json
@@ -1,83 +1,0 @@
-{
-    "id": "3",
-    "slug": "covet",
-    "title": "Covet",
-    "author": "Joanna Virippil",
-    "category": "Editorial",
-    "summary": "An exploration of shared light, stolen identity, and the weight of adornment.",
-    "date": "2026-02-04",
-    "layout": "editorial",
-    "content": [
-        {
-            "type": "image",
-            "url": "/assets/article photos/covet/cereal-1.jpg",
-            "alt": "Title image"
-        },
-        {
-            "type": "text",
-            "value": "Greeting in open arms. The same familiar smile between us. Your light and mine. Woven tightly as threads. No beginning no end. We exchanged secrets, glances. Like they were nothing. But meant everything. Mine feeling more a choker. Dread creeps out from the light."
-        },
-        {
-            "type": "text",
-            "value": "Idiosyncrasy that was left unnoticed. Subtle discord treads the air. If our hearts are truly united, why is your light brighter? Stronger, Sharper, Better."
-        },
-        {
-            "type": "image",
-            "url": "/assets/article photos/covet/11-flick.jpg",
-            "alt": "Jewelry flick caught in motion"
-        },
-        {
-            "type": "text",
-            "value": "You point out my necklace, clutching tightly at your own. Your pristine, polished nails clawing at my nape. You're the same as me aren’t you? Bracelets slip, tangled between wrists."
-        },
-        {
-            "type": "image",
-            "url": "/assets/article photos/covet/8-final.jpg",
-            "alt": "Coat and shoes swapped between friends"
-        },
-        {
-            "type": "image",
-            "url": "/assets/article photos/covet/text.png",
-            "alt": "Narrative graphic"
-        },
-        {
-            "type": "text",
-            "value": "Your coat draped over me, a conquest. My shoes on your feet, stomping out my shadow."
-        },
-        {
-            "type": "image",
-            "url": "/assets/article photos/covet/6-hair.jpg",
-            "alt": "Jewelry being adjusted in hair"
-        },
-        {
-            "type": "text",
-            "value": "I drape your shawl across my shoulders. You slip my earrings into your ears. Each exchange a theft, a gift, a wound. Shadows lengthen in our laughter."
-        },
-        {
-            "type": "text",
-            "value": "We struggle to escape from each other. Left in a disgraceful manner. I tore my gaze away and fell silent. I've become lost in my feelings... how pathetic."
-        },
-        {
-            "type": "image",
-            "url": "/assets/article photos/covet/5-final.jpg"
-        },
-        {
-            "type": "text",
-            "value": "Gilded strangers in borrowed skins. When your jewels glitter on my hands and mine sparkle against your throat, we do not soften. We stare. Jealous still, hungry still, clothed in each other’s stolen light."
-        },
-        {
-            "type": "gallery",
-            "images": [
-                {
-                    "url": "/assets/article photos/covet/3-punch.jpg"
-                },
-                {
-                    "url": "/assets/article photos/covet/4-punch.jpg"
-                },
-                {
-                    "url": "/assets/article photos/covet/DILC.png"
-                }
-            ]
-        }
-    ]
-}
