+++ src/routes/Issues/kismet.jsx
@@ -1,0 +1,10 @@
+import { createFileRoute } from '@tanstack/react-router'
+import KismetPage from '../../components/IssueArticles/kismet/kismetPage'
+
+export const Route = createFileRoute('/Issues/kismet')({
+    component: RouteComponent,
+})
+
+function RouteComponent() {
+    return <KismetPage />
+}
